+++ backend/problems/src/ma1/equations/equations_with_multiple_variable_terms.rs
@@ -193,7 +193,7 @@ fn one_negative_coef_rhs_greater(id: i32, _lang: Language) -> Result<Problem> {
     }))
 }
 
-// 4 - 4x = 8 - 2x
+/// 4 - 4x = 8 - 2x
 /// Absolute difficulty: 4
 /// Relative difficulty: 3
 #[problem]
@@ -238,7 +238,7 @@ fn two_negative_coefs_rhs_greater(id: i32, _lang: Language) -> Result<Problem> {
     }))
 }
 
-// 4 - 2x = 8 - 4x
+/// 4 - 2x = 8 - 4x
 /// Absolute difficulty: 4
 /// Relative difficulty: 3
 #[problem]
@@ -284,7 +284,7 @@ fn two_negative_coefs_lhs_greater(id: i32, _lang: Language) -> Result<Problem> {
     }))
 }
 
-// 4x = 6x - 6
+/// 4x = 6x - 6
 /// Absolute difficulty: 4
 /// Relative difficulty: 4
 #[problem]
@@ -327,7 +327,7 @@ fn positive_coefs_lhs_has_zero(id: i32, _lang: Language) -> Result<Problem> {
     }))
 }
 
-// 6x - 6 = 4x
+/// 6x - 6 = 4x
 /// Absolute difficulty: 4
 /// Relative difficulty: 4
 #[problem]